Enabled importing of first template file.
There were some mismatches between the structure of the template file and the internal structure of the database. These differences have been resolved and the first template can now be imported. Validation/Error handling still need to be implemented.

+++ Database/Relations/Crop.cs
@@ -13,12 +13,12 @@ namespace Database
 
         // For use with Activator.CreateInstance()
         public Crop(
-            int cropId,
+            double cropId,
             string name,
             string notes
         )
         {
-            CropId = cropId;
+            CropId = (int)cropId;
             Name = name;
             Notes = notes;
         }
